Make abstract class & add service principal rotator (#3)
* Implement an abstract base class for rotating secrets

Secret rotation is now done by an abstract base class with concrete
types that implement key-generation steps for specific types of secrets. The
SasTokenSecretRotator class demonstrates how to implement SecretRotator
by specifying A) which types of secrets the rotator applies to, and B)
the steps for generating new secret values.

Secrets themselves are encapsulated in a new Secret class, which is
constructed using a Key Vault name and a secret name. The Secret class
exposes the secret's tags as class properties along with a client for
Key Vault I/O.

Finally, the Event Grid trigger handler now simply works as follows:
1. construct a new Secret object given info from the Event Grid event
2. maintain a mapping of secret types to SecretRotator objects and
select the appropriate SecretRotator object to invoke, logging an error
in the event that no such suitable object is found
3. Invoke the RotateSecret method on the rotator

* add Service Principal rotator

Co-authored-by: Kyle Estes <kestes.cw@mmm.com>